requirements.txt
# requirements.txt
gradio
openai
python-dotenv